# docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CinePair Backend",
    "description": "Real-time co-watching signaling server",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Health check endpoint with room statistics.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready Check",
        "description": "Readiness check endpoint.",
        "operationId": "ready_check_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/api/ice-servers": {
      "get": {
        "summary": "Get Ice Servers",
        "description": "Get STUN/TURN server configuration with ephemeral credentials.",
        "operationId": "get_ice_servers_api_ice_servers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    },
    "/api/rooms": {
      "post": {
        "summary": "Create Room",
        "description": "Create a new room.",
        "operationId": "create_room_api_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_code}/join": {
      "post": {
        "summary": "Join Room",
        "description": "Join a room. Returns 200 (joined), 202 (approval pending), or error.",
        "operationId": "join_room_api_rooms__room_code__join_post",
        "parameters": [
          {
            "name": "room_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomJoinPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_code}": {
      "get": {
        "summary": "Get Room Info",
        "description": "Get public room information.",
        "operationId": "get_room_info_api_rooms__room_code__get",
        "parameters": [
          {
            "name": "room_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_code}/approve": {
      "post": {
        "summary": "Approve Join Request",
        "description": "Approve or deny a join request (admin only).",
        "operationId": "approve_join_request_api_rooms__room_code__approve_post",
        "parameters": [
          {
            "name": "room_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_code}/leave": {
      "post": {
        "summary": "Leave Room",
        "description": "Leave a room (authenticated user).",
        "operationId": "leave_room_api_rooms__room_code__leave_post",
        "parameters": [
          {
            "name": "room_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "description": "Prometheus metrics endpoint.",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApprovalPayload": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 8,
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "code",
          "request_id",
          "approved"
        ],
        "title": "ApprovalPayload",
        "description": "Payload for approving/denying a join request."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoomCreatePayload": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Nickname",
            "description": "Display name of the admin"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Optional room password"
          },
          "require_approval": {
            "type": "boolean",
            "title": "Require Approval",
            "description": "Whether joining requires admin approval",
            "default": true
          },
          "max_users": {
            "type": "integer",
            "maximum": 10,
            "minimum": 2,
            "title": "Max Users",
            "description": "Maximum users allowed in room",
            "default": 2
          },
          "room_expiry_hours": {
            "type": "number",
            "maximum": 168,
            "minimum": 1,
            "title": "Room Expiry Hours",
            "description": "Hours before room expires",
            "default": 24
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "RoomCreatePayload",
        "description": "Payload for creating a new room (via HTTP or Socket.IO)."
      },
      "RoomJoinPayload": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Nickname"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "RoomJoinPayload",
        "description": "Payload for joining a room."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}